backend/enhanced_requirements.txt
# Enhanced Requirements for Federated Learning ViT Pneumonia Classification

# Core Flask and Web
Flask==3.0.3
Flask-Cors==5.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.3

# AI/ML Core
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.46.3
huggingface-hub==0.26.2

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84

# Scientific Computing
numpy==2.1.3
scipy==1.14.1

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Data Processing
pandas==2.2.3

# Networking and HTTP
requests==2.32.3
urllib3==2.2.2

# Federated Learning
# Note: For production, consider PySyft, TensorFlow Federated, or Flower
# flower==1.11.0  # Uncomment for Flower framework

# Security and Cryptography (for FL)
cryptography==43.0.3

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
click==8.1.7
colorama==0.4.6

# Monitoring and Logging
psutil==6.1.0

# Testing (optional)
pytest==8.3.3
pytest-cov==6.0.0

# Development Tools
black==24.10.0
flake8==7.1.1

# Optional: Database support for storing FL metrics
# sqlalchemy==2.0.36
# sqlite3 (built-in)

# Optional: Advanced ML monitoring
# wandb==0.18.6
# tensorboard==2.18.0
